Split classes back up
