+++ tsconfig.json
@@ -11,6 +11,6 @@
         "outDir": "javascripts"
     },
     "include": [
-        "typescript/inline*/rhdp-*.ts"
+        "typescript/**/rhdp-*.ts"
     ]
 }
