+++ typescript/rhdp-search/rhdp-search.ts
@@ -1,4 +1,12 @@
-declare const System: any;
 window.addEventListener('WebComponentsReady', function() {
-    System.import('rhdp-search-app');
+    customElements.define('rhdp-search-sort-page', RHDPSearchSortPage);
+    customElements.define('rhdp-search-query', RHDPSearchQuery);
+    customElements.define('rhdp-search-box', RHDPSearchBox);
+    customElements.define('rhdp-search-result-count', RHDPSearchResultCount);
+    customElements.define('rhdp-search-result', RHDPSearchResult);
+    customElements.define('rhdp-search-results', RHDPSearchResults);
+    customElements.define('rhdp-search-filter-item', RHDPSearchFilterItem);
+    customElements.define('rhdp-search-filter-group', RHDPSearchFilterGroup);
+    customElements.define('rhdp-search-filters', RHDPSearchFilters);
+    customElements.define('rhdp-search-app', RHDPSearchApp);
 });
